Add the service worker file again as push notifications don't work when the app is closed
In Dav.ts replace the push subscription with sending a message to the service worker with the variables necessary for showing notifications, like baseUrl and icon paths. In the service worker, save the values as global variables and show a notification when push is triggered. Additionally handle the notification click event with opening the app.